Rewrite complexity estimates to work with actual work done

+++ src/main/java/com/gmail/snowmanam2/dispenserfill/FillSystemTask.java
@@ -1,5 +1,5 @@
 package com.gmail.snowmanam2.dispenserfill;
 
 public interface FillSystemTask extends Runnable {
-	public int getComplexity();
+	public long getComplexity();
 }
